archive_old_scripts/old_backups/backup_20260130_235425/pyproject.toml
[project]
name = "limo-modern-backend"
version = "0.1.0"
description = "Modern FastAPI backend for Limo app"
authors = [{name = "Arrow Limo"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.112.0",
  "pydantic>=2.8.0",
  "psycopg2-binary>=2.9.9",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.23.8",
]
obs = [
  "sentry-sdk>=2.8.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM","C90","RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
